requirements.txt
openai~=0.27
PyGithub~=1.58
